requirements.txt
nltk == 3.2.2
numpy == 1.12.1
python-Levenshtein == 0.12.0
pycorenlp == 0.3
rdflib == 4.2.2
pandas == 0.20.1